added service implementations classes for enquiry and counsellor service interfaces

+++ src/main/java/in/stproject/dto/DashboardResponseDTO.java
@@ -6,9 +6,9 @@ import lombok.Data;
 public class DashboardResponseDTO {
 
 	
-	private Long totalEnqCnt;
-	private Long openEnqCnt;
-	private Long enrolledEnqCnt;
-	private Long  lostEnqCnt;
+	private Integer totalEnqCnt;
+	private Integer openEnqCnt;
+	private Integer enrolledEnqCnt;
+	private Integer  lostEnqCnt;
 	
 }
